+++ plotly/examples/contour_examples.rs
@@ -1,7 +1,7 @@
-use plotly::{Plot, Contour, Layout};
-use std::f64::consts::PI;
-use plotly::common::{Title, ColorScale, ColorScalePalette};
+use plotly::common::{ColorScale, ColorScalePalette, Title};
 use plotly::contour::Contours;
+use plotly::{Contour, Layout, Plot};
+use std::f64::consts::PI;
 
 fn simple_contour_plot() {
     let n = 200;
@@ -19,13 +19,14 @@ fn simple_contour_plot() {
         let mut row = Vec::<f64>::new();
         for yi in 0..n {
             let radius_squared = x[xi].powf(2.0) + y[yi].powf(2.0);
-            let zv = x[xi].sin() * y[yi].cos() * radius_squared.sin() / (radius_squared + 1.0).log10();
+            let zv =
+                x[xi].sin() * y[yi].cos() * radius_squared.sin() / (radius_squared + 1.0).log10();
             row.push(zv);
         }
         z.push(row);
     }
 
-    let trace = Contour::new(x, y,z);
+    let trace = Contour::new(x, y, z);
     let mut plot = Plot::new();
 
     plot.add_trace(trace);
@@ -33,11 +34,13 @@ fn simple_contour_plot() {
 }
 
 fn colorscale_for_contour_plot() {
-    let z = vec![vec![10.0, 10.625, 12.5, 15.625, 20.0],
+    let z = vec![
+        vec![10.0, 10.625, 12.5, 15.625, 20.0],
         vec![5.625, 6.25, 8.125, 11.25, 15.625],
         vec![2.5, 3.125, 5., 8.125, 12.5],
         vec![0.625, 1.25, 3.125, 6.25, 10.625],
-        vec![0.0, 0.625, 2.5, 5.625, 10.0]];
+        vec![0.0, 0.625, 2.5, 5.625, 10.0],
+    ];
     let trace = Contour::new_z(z).color_scale(ColorScale::Palette(ColorScalePalette::Jet));
 
     let layout = Layout::new().title(Title::new("Colorscale for Contour Plot"));
@@ -48,11 +51,13 @@ fn colorscale_for_contour_plot() {
 }
 
 fn customizing_size_and_range_of_a_contour_plots_contours() {
-    let z = vec![vec![10.0, 10.625, 12.5, 15.625, 20.0],
-                 vec![5.625, 6.25, 8.125, 11.25, 15.625],
-                 vec![2.5, 3.125, 5., 8.125, 12.5],
-                 vec![0.625, 1.25, 3.125, 6.25, 10.625],
-                 vec![0.0, 0.625, 2.5, 5.625, 10.0]];
+    let z = vec![
+        vec![10.0, 10.625, 12.5, 15.625, 20.0],
+        vec![5.625, 6.25, 8.125, 11.25, 15.625],
+        vec![2.5, 3.125, 5., 8.125, 12.5],
+        vec![0.625, 1.25, 3.125, 6.25, 10.625],
+        vec![0.0, 0.625, 2.5, 5.625, 10.0],
+    ];
     let trace = Contour::new_z(z)
         .color_scale(ColorScale::Palette(ColorScalePalette::Jet))
         .auto_contour(false)
@@ -66,11 +71,13 @@ fn customizing_size_and_range_of_a_contour_plots_contours() {
 }
 
 fn customizing_spacing_between_x_and_y_ticks() {
-    let z = vec![vec![10.0, 10.625, 12.5, 15.625, 20.0],
-                 vec![5.625, 6.25, 8.125, 11.25, 15.625],
-                 vec![2.5, 3.125, 5., 8.125, 12.5],
-                 vec![0.625, 1.25, 3.125, 6.25, 10.625],
-                 vec![0.0, 0.625, 2.5, 5.625, 10.0]];
+    let z = vec![
+        vec![10.0, 10.625, 12.5, 15.625, 20.0],
+        vec![5.625, 6.25, 8.125, 11.25, 15.625],
+        vec![2.5, 3.125, 5., 8.125, 12.5],
+        vec![0.625, 1.25, 3.125, 6.25, 10.625],
+        vec![0.0, 0.625, 2.5, 5.625, 10.0],
+    ];
     let trace = Contour::new_z(z)
         .color_scale(ColorScale::Palette(ColorScalePalette::Jet))
         .dx(10.0)