+++ plotly/src/box_plot.rs
@@ -71,6 +71,10 @@ where
     hover_info: Option<HoverInfo>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "hovertemplate")]
     hover_template: Option<Dim<String>>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "xaxis")]
+    x_axis: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "yaxis")]
+    y_axis: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     orientation: Option<Orientation>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "alignmentgroup")]
@@ -143,6 +147,8 @@ where
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             orientation: None,
             alignment_group: None,
             offset_group: None,
@@ -193,6 +199,8 @@ where
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             orientation: None,
             alignment_group: None,
             offset_group: None,
@@ -237,6 +245,8 @@ where
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             orientation: None,
             alignment_group: None,
             offset_group: None,
@@ -330,6 +340,16 @@ where
 
     pub fn hover_template(mut self, hover_template: &str) -> Box<BoxPlot<Y, X>> {
         self.hover_template = Some(Dim::Scalar(hover_template.to_owned()));
+        Box::new(self)
+    }
+
+    pub fn x_axis(mut self, axis: &str) -> Box<BoxPlot<Y, X>> {
+        self.x_axis = Some(axis.to_owned());
+        Box::new(self)
+    }
+
+    pub fn y_axis(mut self, axis: &str) -> Box<BoxPlot<Y, X>> {
+        self.y_axis = Some(axis.to_owned());
         Box::new(self)
     }