+++ plotly/src/heat_map.rs
@@ -34,6 +34,10 @@ where
     hover_info: Option<HoverInfo>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "hovertemplate")]
     hover_template: Option<Dim<String>>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "xaxis")]
+    x_axis: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "yaxis")]
+    y_axis: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "colorbar")]
     color_bar: Option<ColorBar>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "autocolorscale")]
@@ -71,7 +75,9 @@ where
 }
 
 impl<Z> HeatMap<Z, f64, f64>
-where Z: Serialize {
+where
+    Z: Serialize,
+{
     pub fn new_z(z: Vec<Z>) -> Box<HeatMap<Z, f64, f64>> {
         Box::new(HeatMap {
             x: None,
@@ -87,6 +93,8 @@ where Z: Serialize {
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             color_bar: None,
             auto_color_scale: None,
             color_scale: None,
@@ -129,6 +137,8 @@ where
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             color_bar: None,
             auto_color_scale: None,
             color_scale: None,
@@ -193,6 +203,16 @@ where
 
     pub fn hover_template(mut self, hover_template: &str) -> Box<HeatMap<Z, X, Y>> {
         self.hover_template = Some(Dim::Scalar(hover_template.to_owned()));
+        Box::new(self)
+    }
+
+    pub fn x_axis(mut self, axis: &str) -> Box<HeatMap<Z, X, Y>> {
+        self.x_axis = Some(axis.to_owned());
+        Box::new(self)
+    }
+
+    pub fn y_axis(mut self, axis: &str) -> Box<HeatMap<Z, X, Y>> {
+        self.y_axis = Some(axis.to_owned());
         Box::new(self)
     }