+++ plotly/src/scatter.rs
@@ -9,7 +9,6 @@ use crate::private;
 use crate::Trace;
 use serde::Serialize;
 
-
 #[derive(Serialize, Debug)]
 pub struct Scatter<X, Y>
 where
@@ -45,6 +44,10 @@ where
     hover_info: Option<HoverInfo>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "hovertemplate")]
     hover_template: Option<Dim<String>>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "xaxis")]
+    x_axis: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "yaxis")]
+    y_axis: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     orientation: Option<Orientation>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "groupnorm")]
@@ -104,6 +107,8 @@ where
             hover_text: None,
             hover_info: None,
             hover_template: None,
+            x_axis: None,
+            y_axis: None,
             orientation: None,
             group_norm: None,
             stack_group: None,
@@ -213,6 +218,16 @@ where
 
     pub fn hover_template(mut self, hover_template: &str) -> Box<Scatter<X, Y>> {
         self.hover_template = Some(Dim::Scalar(hover_template.to_owned()));
+        Box::new(self)
+    }
+
+    pub fn x_axis(mut self, axis: &str) -> Box<Scatter<X, Y>> {
+        self.x_axis = Some(axis.to_owned());
+        Box::new(self)
+    }
+
+    pub fn y_axis(mut self, axis: &str) -> Box<Scatter<X, Y>> {
+        self.y_axis = Some(axis.to_owned());
         Box::new(self)
     }